+++ lib/hexapdf/layout/text_layouter.rb
@@ -773,6 +773,11 @@ module HexaPDF
             if line.items.empty? && line_fragments.empty?
               # item didn't fit because no more height is available
               next nil if actual_height + item.height > height
+              # item fits but is followed by penalty item that didn't fit
+              if item.width < width_block.call(item.item)
+                too_wide_box = item
+                next nil
+              end
 
               old_height = actual_height
               while item.width > width_block.call(item.item) && actual_height <= height